Create generatemd script
