+++ docs/scripts/generate-llms/index.js
@@ -1,7 +1,7 @@
 import { compileTalksFile } from './compileTalks.js';
+import { generateGenerateMd } from './generatemd/index.js';
 import { generateLlmsEasTxt } from './llms-eas-txt.js';
 import { generateLlmsFullTxt } from './llms-full-txt.js';
-import { generatePerPageMd } from './per-page/index.js';
 import { generateLlmsSdkTxt } from './llms-sdk.js';
 import { generateLlmsTxt } from './llms-txt.js';
 
@@ -12,5 +12,5 @@ Promise.allSettled([
   await generateLlmsEasTxt(),
   await generateLlmsFullTxt(),
   await generateLlmsTxt(),
-  await generatePerPageMd(),
+  await generateGenerateMd(),
 ]).catch(console.error);
